minor - rename ceed-jit.c as ceed-jit-tools.c
